backend/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
jieba>=0.42.0
snownlp>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
requests>=2.31.0
openai>=1.0.0
# transformers>=4.35.0  # 暂时注释，因为需要PyTorch
# torch>=2.0.0          # 暂时注释，Python 3.13兼容性问题
# sentence-transformers>=2.2.0  # 暂时注释，因为需要PyTorch
# 替代方案：使用轻量级文本处理库
scikit-learn>=1.3.0
textblob>=0.17.0 